Restructured project folder heirarchy before begging data restructure
