--- app/src/main/java/com/example/subget/MyData.kt
@@ -1,3 +1,0 @@
-package com.example.subget
-
-class MyData : ArrayList<MyDataItem>()
