+++ app/src/main/java/com/example/subget/app_data/models/AllListings.kt
@@ -1,0 +1,3 @@
+package com.example.subget.app_data.models
+
+class AllListings (val results : List<Listing>)
