+++ app/src/main/java/com/example/subget/ui/listing_detail/DetailsFragment.kt
@@ -1,19 +1,16 @@
-package com.example.subget.home
+package com.example.subget.ui.listing_detail
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.bumptech.glide.Glide
-import com.example.subget.R
 import com.example.subget.databinding.FragmentDetailsBinding
-import com.example.subget.databinding.FragmentHomeBinding
 
 
-class HomeFragment : Fragment() {
+class DetailsFragment : Fragment() {
 
-    private var _binding: FragmentHomeBinding? = null
+    private var _binding: FragmentDetailsBinding? = null
     private val binding get() = _binding!!
 
 
@@ -22,7 +19,7 @@ class HomeFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
 
-        _binding = FragmentHomeBinding.inflate(inflater, container, false)
+        _binding = FragmentDetailsBinding.inflate(inflater, container, false)
 
         return binding.root
     }
