+++ app/src/main/java/com/example/subget/utils/Constants.kt
@@ -1,0 +1,8 @@
+package com.example.subget.utils
+
+class Constants {
+
+    companion object {
+        const val BASE_URL = "https://subget.herokuapp.com/"
+    }
+}
