+++ app/src/main/res/layout/fragment_favorites.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".favorites.FavoritesFragment">
+    tools:context=".ui.favorites.FavoritesFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
