+++ app/src/main/res/layout/fragment_listings.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".listings.ListingsFragment">
+    tools:context=".ui.listings.ListingsFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
